Fix previous commit (was a big mess, sorry). Also restores the correct behaviour of mmaps on pre-cata versions (a typo?).
