+++ Movemap-Generator/CMakeLists.txt
@@ -56,7 +56,6 @@ set(SOURCES
     ${CMAKE_CURRENT_SOURCE_DIR}/IntermediateValues.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/generator.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/MapBuilder.cpp
-    ${CMAKE_CURRENT_SOURCE_DIR}/MapBuilderTask.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/DelayExecutor.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/TerrainBuilder.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/VMapExtensions.cpp
